Wrote Experiment class
The Experiment class creates and manages directories for saving all data associated with an experiment. 
Right now, the default database is sqlite, and this class will create a new folder for each experiment and then create sqlite files in there as needed.
The TableSource class has also been updated so that its query method returns a DBSource, allowing one to iterate through queries using a Source.