Wrote open method and began work on load_experiment function
The open method can be used to return a handle to a file in the experiment's directory (or a string containing its path). This can be used to save arbitrary files associated with an experiment in the same folder.
load_experiment is a function that returns an Experiment object corresponding to an existing directory. This can be used to regenerate plots or continue an experiment as desired.