Updates to database and factory modules.
The TableSource can now infer columns based on its table, so that you don't have to specify column names upon construction (you can optionally specify them in order to push to a subset of the columns).

The factory module is being retooled to make SQL the default method for storing and reading parameters/metrics for hyperparameter optimization, because that is ultimately cleaner than the current approach, which is more document oriented in nature. Alternatively, we could maintain this approach and implement a document-DB backend, but that would add complexity due to multiple database types being used simultaneously.
This is a work in progress at the moment.