Added RepeaterSource
A RepeaterSource takes an iterable Source and causes it to repeat multiple times before raising StopIteration.
For example, if an iterator has 10 elements, you can make the RepeaterSource loop through that iterator 5 times to get a total of 50 samples. This can be useful if you want to oversample a dataset.