Fixed bugs involving filter method
The filter method was throwing errors due to incorrectly inferring column names for certain queries. I don't fully understand what the best way to infer the column names for a query/table/sqlobject is at the moment, but I got it to work for now.
When you use the query method to ask for a subset of columns in a table, you have to keep track of which columns are being used, which we now do. Previously, we were using table object associated with a DBSource for a query to do this, and that would result in errors because the table object has all the columns.