Finished modifications to Pipeline module.
The syntax for pipes has been simplified. Now, in order to create a pipe and link to an input, you can just provide the input as a positional argument. Eg. :

A = pipe_class(B) 

will initialize class pipe_class with pipe B as its input. 

This syntax change is backwards incompatible with previous iterations.