Models now work as envisioned.
Models can function as Pipes, Junctions, and Modules. They can call on the funcitonality of any of these entities in order to perform arbitrarily complex tasks.

For example, you can have a Model that takes as junction-input a Junction that randomly calls upon one of multiple other Models. 
You can make this Model part of a pipeline that includes preprocessing and other Models, and you can link the parameters of this Model to another Model.

model_tests.py demonstrates a few of these instances.