Code Clean Up
Removed old code that was commented out.

I also fixed a potential issue with ModelFromModule class. Since it inherits from the provided class and the Model class, there is some ambiguity in which parent is used for the __call__ method. I explicitly coded in to use the Model class' version, since that one incorporates the recursion that enables models to be stacked in a Pipeline and feed their output into each other.