Code refactoring. (#63)
* Changed Junction class to call it's inputs 'components' instead of 'junction_inputs'

I also updated the Model class so that it has one set of components which encapsulates these Junction components, tensor Parameters, and PyTorch Modules.

* Refactored Pipe Class to call its inputs 'input' instead of 'input_pipe'

There is a built-in function in Python called input, but this does not conflict with anything here.

* Code Clean Up

Removed old code that was commented out.

I also fixed a potential issue with ModelFromModule class. Since it inherits from the provided class and the Model class, there is some ambiguity in which parent is used for the __call__ method. I explicitly coded in to use the Model class' version, since that one incorporates the recursion that enables models to be stacked in a Pipeline and feed their output into each other.