Updates to database module.
If you consider methods to be distinct from an object's attributes, then sessionscope and transaction scope are decoupled here. The data manipulation methods do not manipulate the session in any way (they don't create a new session, commit, or rollback). Those actions take place outside of the method and within the application (you have to explicitly call commit). 
The only difference is that the DBPipe and TablePipes also have their own session attributes. This is there as a convenience, and the session can be reset or replaced with a user provided session by calling the reset_session method. This should satisfy the requirement to keep the session life-cycle separate from the data manipulation logic.