Fixed Variance
Variance was wrong because of a stupid mistake. However, the algorithm to compute variance is still numerically unstable.
I also managed to get recursive method calls to work with Models (although this needs to be tested). If a Model has inputs, you can now call methods/attributes on those inputs like any other Pipe. Previously this was not working because of method resolution errors.