Added to_x methods to Message for serialization.
to_x works by converting all columns of a Message to dataframe and then calling the to_x method on the dataframe. 
This will be used to implement save methods for Models. The Model will convert it's state_dict to a Message of length 1 and then use this serialization to save those values.