Added @recursive decorator.
By adding this decorator to a method, you can indicate that the method call should be recursive. This means that the method will be called on the input as well. This can be useful for configurational methods. For example, if I have a method called 'reset', which is defined for multiple pipes, this decorator can ensure that reset is called for all pipes in the pipeline when it is called downstream.

Additionally, you can specify accumulate=True to indicate that these recursive calls should be composed. For example, if I have a method f that increments its argument by 1, then that incremented argument will be the argument to each successive downstream call.