Progress on model linking and saving.
The idea is that if you want to use an attribute of a Model as a component of another Model, you need to implicitly link these Models together. This is different from just setting model1.x = model2.x, because if you do that, then model1 has no knowledge of model2's existence, and this makes it difficult to keep track of where a variable came from. This is important for recursive operations such as saving and loading, because we want that when model1 is saved, model1.x is treated as a part of model2 and is saved under model2 instead of under model1.