Introduced Component_Mapper class
This class will keep track of subcomponents and state. We will use this to abstract out the tracking of components. This gives us a simpler way to track components that are part of another object while retaining a familiar syntax. This is important, because for example, if we want that m1.x = m2.x to mean that the x attribute of m1 is actually the x attribute of m2, then this gives us a way to do that. Simply setting m1.x = m2.x will assign by value, so when m2.x changes, m1.x will not change. Also, m1 will have no knowledge of m2. We can now do m1.x = (m2, 'x') to indicate this preference.