Scaffold object is complete.
A scaffold is responsible for tracking the state of objects attached to it. Pipes, Junctions, and Models all implement a 'get_state' method that returns a serialized representation of their internal and external state, and the scaffold uses this method to save and load states.
The identity of objects is kept track of via the name assigned to them when attaching. For example, if you call scaffold.attach('a', A), then the object A will be attached with the name 'a'. It will save and load A's state under that name (eg. 'a.json'). 
Note that the Scaffold does not track the connectivities between objects in a pipeline. This is a feature that can be added in the future, but is not essential to save and load pipelines. As long as you can assign a name to every component of a pipeline and keep those names consistent, you can maintain and manipulate connectivities as you desire.