Nonlinear regression and model selection examples complete.
The model selection example demonstrates using factory to compare different polynomial models. It trains and computes metrics for these models and then plots an animation comparing the different saved models.

I will next work on a database example that modifies the previous two tutorials so that training data is read from a database, and the factory writes its parameters to a database as well.