Added new save and load methods for Messages using a binary format.
This introduces a new serialization format designed to retain the tensor/dataframe structure across saving and loading. Currently, if you save as CSV or JSON or some other format, the information about which columns are tensors will be lost.
This format is essentially a tarfile containing a parquet compressed representation of the dataframe and pickled representations of the tensors (via torch.save).